requirements.txt
aiogram==3.4.1
python-dotenv==1.0.1
APScheduler==3.10.4
aiosqlite==0.20.0
openpyxl==3.1.2
reportlab==4.1.0
# aiogram 3.4.1 -> pydantic >=2.4.1,<2.6
pydantic==2.5.3
